+++ Runtime/Authoring/Hybrid/NetCodeClientAndServerSettings.cs
@@ -73,11 +73,35 @@ namespace Unity.NetCode.Hybrid
         {
             if (AssetDatabase.IsAssetImportWorkerProcess())
                 return;
-            Save(true);
             ((IEntitiesPlayerSettings)this).RegisterCustomDependency();
+            Save(true);
+            AssetDatabase.Refresh();
         }
 
-        private void OnDisable() => Save();
+#if UNITY_2023_2_OR_NEWER
+        private void OnEnable()
+        {
+            ((IEntitiesPlayerSettings)this).RegisterCustomDependency();
+        }
+#endif
+        private void OnDisable()
+        {
+#if !UNITY_2023_2_OR_NEWER
+            Save();
+#else
+            //But the depedency is going to be update when the scriptable is re-enabled.
+            if (AssetDatabase.IsAssetImportWorkerProcess())
+                return;
+            Save(true);
+            //This safeguard is necessary because the RegisterCustomDependency throw exceptions
+            //if this is called when the editor is refreshing the database.
+            if(!EditorApplication.isUpdating)
+            {
+                ((IEntitiesPlayerSettings)this).RegisterCustomDependency();
+                AssetDatabase.Refresh();
+            }
+#endif
+        }
     }
 }
 #endif