+++ Runtime/Authoring/Hybrid/NetCodeClientSettings.cs
@@ -89,10 +89,33 @@ namespace Unity.NetCode.Hybrid
         {
             if (AssetDatabase.IsAssetImportWorkerProcess())
                 return;
+            ((IEntitiesPlayerSettings)this).RegisterCustomDependency();
             Save(true);
+            AssetDatabase.Refresh();
+        }
+#if UNITY_2023_2_OR_NEWER
+        private void OnEnable()
+        {
             ((IEntitiesPlayerSettings)this).RegisterCustomDependency();
         }
-        private void OnDisable() { Save(); }
+#endif
+        private void OnDisable()
+        {
+#if !UNITY_2023_2_OR_NEWER
+            Save();
+#else
+            //But the depedency is going to be update when the scriptable is re-enabled.
+            if (AssetDatabase.IsAssetImportWorkerProcess())
+                return;
+            //This safeguard is necessary because the RegisterCustomDependency throw exceptions
+            //if this is called when the editor is refreshing the database.
+            if(!EditorApplication.isUpdating)
+            {
+                ((IEntitiesPlayerSettings)this).RegisterCustomDependency();
+                AssetDatabase.Refresh();
+            }
+#endif
+        }
     }
 
     internal class ClientSettings : DotsPlayerSettingsProvider